Move env var loading into a separate file.
This is because when using emacs native-comp, env var loading must be
disabled on first startup otherwise the native compilation process will
be disrupted by a missing executable. Moving it into a separate file
makes it so that I can just comment out the `load-file` line.

I believe this could be remedied by running the gen-env-file.el script
from a plain emacs launched from PowerShell as that would have both
the emacs and shell environment.